test #if HAS_SPAN in existing file
