+++ NBitcoin.Tests/DeterministicSignatureTests.cs
@@ -250,6 +250,36 @@ namespace NBitcoin.Tests
 
 			return key;
 		}
+
+		[Fact]
+		[Trait("UnitTest", "UnitTest")]
+		public void ECDSASignatureIsOnlyBip66DER()
+		{
+			var invalidBIP66DER = HexToByteArray("302402107777777777777777777777777777777702108777777777777777777777777777777701");
+			// Encodes:
+			// SEQUENCE {
+			//	 INTEGER 0x77777777777777777777777777777777,
+			//   INTEGER 0x87777777777777777777777777777777 // -1 exponent :(
+			// }
+			// BOOLEAN FALSE
+			var coercedSignature = new ECDSASignature(invalidBIP66DER);
+
+			var validBIP66DER = HexToByteArray("302502107777777777777777777777777777777702110087777777777777777777777777777777");
+			// Should coerce to this valid DER encoding:
+			// SEQUENCE {
+			//	 INTEGER 0x77777777777777777777777777777777,
+			//   INTEGER 0x0087777777777777777777777777777777 // +1 exponent :)
+			// }
+			Assert.True(coercedSignature.ToDER().SequenceEqual(validBIP66DER));
+		}
+
+		private static byte[] HexToByteArray(string hex)
+		{
+			byte[] bytes = new byte[hex.Length / 2];
+			for (int i = 0; i < hex.Length; i += 2)
+				bytes[i / 2] = Convert.ToByte(hex.Substring(i, 2), 16);
+			return bytes;
+		}
 #endif
 
 		[Fact]
